Correct ship attack level typing

+++ src/rotp/model/empires/ShipView.java
@@ -158,7 +158,7 @@ public class ShipView implements Base,Serializable {
     }
     private void scanComputer() {
         computer = design.computer();
-        attackLevel = (int) design.attackLevel()+empire.shipAttackBonus();
+        attackLevel = design.attackLevel()+empire.shipAttackBonus();
         computerKnown = true;
         attackLevelKnown = true;
     }
